[bazel] Update example workspace files with dependencies
Currently, the example workspace files do not work properly. The
llvm_disable_optional_support_deps option was removed in 0d8a320
and zlib and zstd have been needed-by-default dependencies for a while,
so it would make sense to show definitions in the example bazel
workspace files, especially given the version sensitivity of the zlib-ng
build. This patch removes the use of dated build config flags and adds
dependency definitions to both of the example workspaces.

Reviewed By: aaronmondal

Differential Revision: https://reviews.llvm.org/D156654