[libc++] Reland CI module improvements.
Revert "Revert #76246 and #76083"

This reverts commit ca16771f99c8bf12a08deb3ae483c61cac756785.

Adds a small fix that should properly disable the tests on Windows.
Unfortunately the original poster has not provided feedback and the
original patch did not fail in the LLVM CI infrastructure.

Modules are known to fail on Windows due to non compliance of the
C library. Currently not having this patch prevents testing on other
platforms.