[libc++] Revert new test for integral_constant that breaks CI
This commit reverts 0606c7a31b and 13db0eaf, which added a test to
ensure that type traits are derived from integral_constant. While that
is a fine test to add, the commit didn't go through a PR and as a result
it looks like some of our CI has been broken by it.

This should be an uncontroversial change, but let's re-land it via a PR
to get our usual CI coverage.